Move types to `/types/frontend.ts`
