+++ dotcom-rendering/fixtures/manual/nav-world.ts
@@ -1,4 +1,4 @@
-import type { FENavType } from '../../src/types/types';
+import type { FENavType } from '../../src/types/frontend';
 
 /** `xhs https://www.theguardian.com/world.json | jq .config.nav | pbcopy` */
 export const navWorld = {
