+++ dotcom-rendering/src/lib/decideTheme.ts
@@ -1,5 +1,5 @@
 import { ArticleSpecial, type ArticleTheme, Pillar } from '@guardian/libs';
-import type { FEFormat } from '../types/types';
+import type { FEFormat } from '../types/frontend';
 
 export const decideTheme = ({ theme }: Partial<FEFormat>): ArticleTheme => {
 	switch (theme) {
