+++ dotcom-rendering/src/model/buildLightboxImages.ts
@@ -6,7 +6,7 @@ import type {
 	ImageBlockElement,
 	ImageForLightbox,
 } from '../types/content';
-import type { FEFormat } from '../types/types';
+import type { FEFormat } from '../types/frontend';
 import {
 	getCartoonImageForLightbox,
 	isCartoon,
