+++ dotcom-rendering/src/types/article.amp.tsx
@@ -1,7 +1,7 @@
 import type { EditionId } from '../lib/edition';
-import type { FEFormat, FELinkType, LegacyPillar } from '../types/types';
 import type { CommercialProperties } from './commercial';
 import type { FEElement } from './content';
+import type { FEFormat, FELinkType, LegacyPillar } from './frontend';
 import type { TagType } from './tag';
 
 // This is a subset of FEArticleType for use in AMP and as a result there needs to be parity between the types of shared fields.
